% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_ip.R
\name{change_ip}
\alias{change_ip}
\title{Run OpenVPN and Change IP Address}
\usage{
change_ip(
  config_path = "~/.ovpn",
  exposed_ip,
  country,
  ignore_files = NULL,
  verbose = TRUE
)
}
\arguments{
\item{config_path}{The path to the folder containing server configuration files.}

\item{exposed_ip}{User non-protected public IPv4 address obtained with \code{get_ip()}.}

\item{country}{\link{optional} The ISO-2 code of the country to pick up a server.}

\item{ignore_files}{\link{optional} TVector of servers configuration file not to be connected.}

\item{verbose}{A boolean. If TRUE, prints connexion informations.}
}
\value{
The selected server.
}
\description{
This function changes your IP Address by connecting to a randomly
selected third-party server using OpenVPN software.
}
\examples{
\dontrun{
change_ip()
}
}
\author{
Nicolas CASAJUS, \email{nicolas.casajus@gmail.com}
}
