% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-system.R
\name{check_system}
\alias{check_system}
\title{Check user system before using the package}
\usage{
check_system(verbose = TRUE)
}
\arguments{
\item{verbose}{a \code{logical}. If \code{TRUE}, displays information.}
}
\value{
No return value. An error is raised if user system is not correctly
set.
}
\description{
Checks user system before using the package. In particular, checks if:
\itemize{
\item the OS is based on Unix (no support for Windows)
\item the user password is correctly stored
\item the system has an Internet connection
\item R is running outside RStudio Desktop
\item the software OpenVPN is installed and accessible
\item VPN configuration files are stored locally
\item the software Docker is installed and running
\item the Docker image Selenium has been downloaded
}
}
\examples{
\dontrun{
check_system()
}
}
