% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/close_vpn.R
\name{close_vpn}
\alias{close_vpn}
\title{Disconnect from VPN and close OpenVPN Daemon}
\usage{
close_vpn(verbose = TRUE)
}
\arguments{
\item{verbose}{A boolean. If TRUE, prints informations.}
}
\value{
Nothing.
}
\description{
This function disconnects from a VPN server and closes the OpenVPN Daemon.
}
\examples{
\dontrun{
close_vpn()
}
}
\author{
Nicolas CASAJUS, \email{nicolas.casajus@gmail.com}
}
