% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escape_url2.R
\name{escape_url2}
\alias{escape_url2}
\title{Escape Whitespace in URL}
\usage{
escape_url2(x)
}
\arguments{
\item{x}{A string to escape whitespaces.}
}
\value{
A escaped string.
}
\description{
This function escapes all whitespaces in a URL.
}
\examples{
\dontrun{
escape_url2("Canis lupus lupus")
escape_url2("https://google.com/canis lupus")
}
}
\author{
Nicolas CASAJUS, \email{nicolas.casajus@gmail.com}
}
