% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-countries.R
\name{get_countries}
\alias{get_countries}
\title{Get the ISO-2 code of countries}
\usage{
get_countries(pattern = NULL)
}
\arguments{
\item{pattern}{a \code{character} of length 1. A regular expression to be matched
(optional).}
}
\value{
A 2-column \code{data.frame} with:
\itemize{
\item \code{country}: the country name
\item \code{iso_2}: the country ISO-2 code
}
}
\description{
Returns names and ISO-2 codes for 57 World countries (listed from NordVPN
servers). User can search for specific patterns (see examples below).
}
\examples{
get_countries()
get_countries(pattern = "Fr")
}
