% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ip.R
\name{get_ip}
\alias{get_ip}
\title{Get System Public IP Address}
\usage{
get_ip()
}
\value{
The IPv4 address (character of length 1)
}
\description{
This function retrieves the public IP address of the system using the ipify API (\url{https://api.ipify.org}).
}
\examples{
\dontrun{
get_ip()
}
}
\author{
Nicolas CASAJUS, \email{nicolas.casajus@gmail.com}
}
