% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_languages.R
\name{get_languages}
\alias{get_languages}
\title{Get ISO-2 Code of Google Scholar Languages}
\usage{
get_languages(pattern)
}
\arguments{
\item{pattern}{A character of length 1 containing a regular expression to be matched (optional).}
}
\value{
A 2-columns data frame with:
\itemize{
\item language: the language name
\item iso_2: the language ISO-2 code
}
}
\description{
This function returns names and ISO-2 codes for 43 World
languages used by Google Scholar. User can search for specific patterns
(see examples below).
}
\examples{
\dontrun{
get_languages()
get_languages(pattern = "spain")
get_languages(pattern = "Sp")
get_languages(pattern = "^Sp")
get_languages(pattern = "^Sp")$iso_2
}
}
\author{
Nicolas CASAJUS, \email{nicolas.casajus@gmail.com}
}
