% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ua.R
\name{get_ua}
\alias{get_ua}
\title{Check How the User Agent is Detected}
\usage{
get_ua(agent, verbose = TRUE)
}
\arguments{
\item{agent}{A character of length 1 specifying a user agent.}

\item{verbose}{A boolean. If TRUE, the new user agent is printed.}
}
\description{
This function checks how a user agent is detected by a server using
the What's My UA API (\url{https://www.whatsmyua.info/api/v1/ua}).
}
\examples{
\dontrun{
get_ua()

ua <- change_ua()
get_ua(agent = ua)

ua <- change_ua("Mozilla/5.0 (Linux x86_64) Gecko/20100101 Firefox/72.0")
get_ua(agent = ua)
}
}
\author{
Nicolas CASAJUS, \email{nicolas.casajus@gmail.com}
}
