% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpack-package.R
\docType{package}
\name{gpack-package}
\alias{gpack}
\alias{gpack-package}
\title{gpack: Tools for webscraping Google Services (Scholar, Pictures, Trends, Search)}
\description{
This package provides tools for webscraping Google Services (Scholar, Pictures, Trends, Search) using OpenVPN to get around IP bans. Webscrap can be performed using RSelenium or classical webscraping (\code{rvest} + \code{httr}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ahasverus/gpack}
  \item Report bugs at \url{https://github.com/ahasverus/gpack/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Casajus \email{nicolas.casajus@gmail.com}

}
\keyword{internal}
