% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_connected.R
\name{is_connected}
\alias{is_connected}
\title{Check for Internet Connexion}
\usage{
is_connected()
}
\value{
Nothing. An error is generated if your system is disconnected from Internet.
}
\description{
This function checks if the sustem is connected to Internet.
R/RStudio must be running under super-user do (sudo).
}
\examples{
\dontrun{
is_connected()
}
}
\author{
Nicolas CASAJUS, \email{nicolas.casajus@gmail.com}
}
