% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_sudo.R
\name{is_sudo}
\alias{is_sudo}
\title{Check for Super User (sudo)}
\usage{
is_sudo()
}
\value{
Nothing. An error is generated if this prerequisite is not met.
}
\description{
This function checks if R/RStudio is running under super-user do (sudo).
This is required to control OpenVPN and change IP address.
}
\examples{
\dontrun{
is_sudo()
}
}
\author{
Nicolas CASAJUS, \email{nicolas.casajus@gmail.com}
}
