% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_unix.R
\name{is_unix}
\alias{is_unix}
\title{Check for Unix System}
\usage{
is_unix()
}
\value{
Nothing. An error is generated if this prerequisite is not met.
}
\description{
This function checks if the system is running under Unix
(GNU/Linux and macOS). The package is not develop for others platforms.
}
\examples{
\dontrun{
is_unix()
}
}
\author{
Nicolas CASAJUS, \email{nicolas.casajus@gmail.com}
}
