% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrap_gimages.R
\name{scrap_gimages}
\alias{scrap_gimages}
\title{Download Images from Google Images}
\usage{
scrap_gimages(
  search_terms,
  start = 0,
  n_max = NULL,
  rs_driver = NULL,
  openvpn = TRUE,
  config_path = "~/.ovpn",
  ovpn_country,
  agent = TRUE,
  sleep = 1,
  verbose = TRUE,
  output_path = ".",
  exposed_ip = NULL
)
}
\arguments{
\item{search_terms}{A character vector of terms to search publications for.}

\item{start}{A numeric specifying the number of the first results from which the results are extracted.}

\item{n_max}{A numeric specifying the number of results to extract.}

\item{rs_driver}{An RSelenium server.}

\item{openvpn}{A boolean. If TRUE, public IP address will be randomly changed.}

\item{config_path}{The path to the folder containing server configuration files.}

\item{ovpn_country}{The ISO-2 code of the country to pick up a server. Use \code{get_countries()} to get a list.}

\item{agent}{A boolean. If TRUE, browser user agent will be randomly changed.}

\item{sleep}{The time interval (in seconds) between two sub-requests.}

\item{verbose}{A boolean. If TRUE, connexion and webscraping informations are printing.}

\item{output_path}{The path to the folder to save data.}

\item{exposed_ip}{The unprotected IPv4 address.}
}
\value{
The RSelenium server
}
\description{
This function sends a request to Google Scholar service and
retrieves results (title, authors, source and year of publications,
and the total number of citations). As no API is provided by Google Scholar
(except the one for authors with a Google Scholar ID), this function webscraps
the service using \code{RSelenium} or the packages \code{rvest} and \code{httr}.
To bypass Google IP bans, user is strongly encouraged to set the arguments
\code{openvpn} and \code{agent} as \code{TRUE}. In this way, every 10-20
sub-requests your IP and your user agent will be automatically and randomly changed.
}
\examples{
\dontrun{
scrap_gimages()
}
}
\author{
Nicolas CASAJUS, \email{nicolas.casajus@gmail.com}
}
