% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stick.R
\name{stick}
\alias{stick}
\title{Paste in a Single Line a Multi-lines String}
\usage{
stick(..., indent = "")
}
\arguments{
\item{...}{String to collapse.}

\item{indent}{Character to add before the string.}
}
\value{
A character.
}
\description{
This function pastes in a single line a multi-lines string.
}
\examples{
\dontrun{
x <-
 "
   Hello
   world!
 "

cat(x)
cat(stick(x))

message(stick(x))
message(stick(x, indent = ">>> "))

usethis::ui_done(stick(x))
usethis::ui_done(stick(x, indent = " "))
}
}
\author{
Nicolas CASAJUS, \email{nicolas.casajus@gmail.com}
}
