% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unix_password.R
\name{unix_password}
\alias{unix_password}
\title{Get the UNIX Password Stored in .Renviron}
\usage{
unix_password()
}
\value{
The UNIX password value.
}
\description{
This function gets the UNIX password stored in .Renviron under the key \code{UNIX_PASSWD}.
Run \code{usethis::edit_r_environ()} and add the following line \code{UNIX_PASSWD='*********'} (make sure to
replace \code{*********} by your password).
}
\examples{
\dontrun{
unix_password()
}
}
\author{
Nicolas CASAJUS, \email{nicolas.casajus@gmail.com}
}
