% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_ip.R
\name{change_ip}
\alias{change_ip}
\title{Run OpenVPN and Change IP Address}
\usage{
change_ip(config_path = "~/.ovpn", exposed_ip, country,
  ignore_files = NULL, verbose = TRUE)
}
\arguments{
\item{config_path}{The path to the folder containing server configuration files.}

\item{exposed_ip}{User unprotected public IPv4 address obtained with \code{close_vpn()}.}

\item{country}{(optional) The ISO-2 code of the country (>= 1) from which a VPN server will be pick up .}

\item{ignore_files}{(optional) Vector of servers configuration file not to be connected.}

\item{verbose}{A boolean. If TRUE, prints connexion informations.}
}
\value{
The selected server.
}
\description{
This function changes your IP Address by connecting to a randomly
selected third-party server using OpenVPN software.
}
\examples{
\dontrun{
close_vpn()
ip <- get_ip()
change_ip(exposed_ip = get_ip())

close_vpn()
change_ip(exposed_ip = get_ip(), country = "es")
}
}
\author{
Nicolas CASAJUS, \email{nicolas.casajus@gmail.com}
}
