% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_ua.R
\name{change_ua}
\alias{change_ua}
\title{Get a Random User Agents}
\usage{
change_ua()
}
\value{
The selected user agent.
}
\description{
This function randomly selects a user agent among a list of 81 obtained from \url{https://techblog.willshouse.com/2012/01/03/most-common-user-agents/} (Last Updated: Fri, 31 Jan 2020).
}
\examples{
\dontrun{
change_ua()
}
}
\author{
Nicolas CASAJUS, \email{nicolas.casajus@gmail.com}
}
