% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_countries.R
\name{get_countries}
\alias{get_countries}
\title{Get ISO-2 Code of Countries}
\usage{
get_countries(pattern)
}
\arguments{
\item{pattern}{A character of length 1 containing a regular expression to be matched (optional).}
}
\value{
A 2-columns data frame with:
\itemize{
\item country: the country name
\item iso_2: the country ISO-2 code
}
}
\description{
This function returns names and ISO-2 codes for 57 World
countries (listed from NordVPN servers). User can search for specific patterns
(see examples below).
}
\examples{
\dontrun{
get_countries()
get_countries(pattern = "Fre")
get_countries(pattern = "Fr")
get_countries(pattern = "^Fr")
get_countries(pattern = "^Fr")$iso_2
}
}
\author{
Nicolas CASAJUS, \email{nicolas.casajus@gmail.com}
}
