% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-languages.R
\name{get_languages}
\alias{get_languages}
\title{Get the ISO-2 code of G**gle Scholar languages}
\usage{
get_languages(pattern = NULL)
}
\arguments{
\item{pattern}{a \code{character} of length 1. A regular expression to be matched
(optional).}
}
\value{
A 2-column \code{data.frame} with:
\itemize{
\item \code{language}: the language name
\item \code{iso_2}: the ISO-2 code
}
}
\description{
Returns names and ISO-2 codes for 43 World languages supported by
\strong{Scholar}. User can search for specific patterns (see examples
below).
}
\examples{
get_languages()
get_languages(pattern = "^Sp")
}
