% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpack-package.R
\docType{package}
\name{gpack-package}
\alias{gpack}
\alias{gpack-package}
\title{gpack: Webscrap Google Services (Scholar, Pictures, Trends, Search)}
\description{
This package provides tools for webscraping Google Services (Scholar, Pictures, Trends, Search) using OpenVPN to get around IP bans. Webscrap is performed by RSelenium.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ahasverus/gpack}
  \item Report bugs at \url{https://github.com/ahasverus/gpack/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Casajus \email{nicolas.casajus@gmail.com} (\href{https://orcid.org/0000-0002-5537-5294}{ORCID}) [copyright holder]

}
\keyword{internal}
