% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrap_gscholar.R
\name{scrap_gscholar}
\alias{scrap_gscholar}
\title{Retrieve Data from Google Scholar}
\usage{
scrap_gscholar(
  search_terms,
  exact = TRUE,
  exclude_terms = NULL,
  search_author = NULL,
  search_source = NULL,
  where = NULL,
  years = NULL,
  lang = NULL,
  start = 0,
  include_patents = FALSE,
  include_citations = FALSE,
  selenium = FALSE,
  browser = "firefox",
  openvpn = TRUE,
  config_path = "~/.ovpn",
  exposed_ip,
  ovpn_country,
  user_agent = TRUE,
  sleep = 1,
  verbose = TRUE
)
}
\arguments{
\item{search_terms}{A character vector of terms to search publications for.}

\item{exact}{A boolen. If TRUE, searchs for the exact terms, otherwise searchs at least one of the terms.}

\item{exclude_terms}{A character vector of terms to exclude from the search.}

\item{search_author}{A character vector of authors to search for.}

\item{search_source}{A character vector of publication sources to search for.}

\item{where}{Search in the whole document ('any') or only in the title ('title').}

\item{years}{A vector of 1 or 2 years specifying the temporal extent of the search.}

\item{lang}{The ISO-2 code of the language to search for. Use \code{get_languages()} to get a list.}

\item{start}{A numeric specifying the number of the first results from which the results are extracted.}

\item{include_patents}{A boolean. If TRUE, patents are included in the search results.}

\item{include_citations}{A boolean. If TRUE, citations are included in the search results.}

\item{selenium}{A boolean. If TRUE, webscraping is performed with the Selenium technology.}

\item{browser}{The browser name ('chrome' or 'firefox'). Ignored if \code{selenium = FALSE}.}

\item{openvpn}{A boolean. If TRUE, public IP address will be randomly changed.}

\item{config_path}{The path to the folder containing server configuration files.}

\item{exposed_ip}{User non-protected public IPv4 address obtained with \code{get_ip()}.}

\item{ovpn_country}{The ISO-2 code of the country to pick up a server. Use \code{get_countries()} to get a list.}

\item{user_agent}{A boolean. If TRUE, browser user agent will be randomly changed.}

\item{sleep}{The time interval (in seconds) between two sub-requests.}

\item{verbose}{A boolean. If TRUE, connexion and webscraping informations are printing.}
}
\value{
A 7-columns data frame with:
\itemize{
\item query: the query terms
\item gsid: the publication Google Scholar ID
\item title: the publication title
\item authors: the publication authors
\item source: the publication source
\item year: the publication year
\item citation: the number of citations of the publication
}
}
\description{
This function sends a request to Google Scholar service and
retrieves results (title, authors, source and year of publications,
and the total number of citations). As no API is provided by Google Scholar
(except the one for authors with a Google Scholar ID), this function webscraps
the service using \code{RSelenium} or the packages \code{rvest} and \code{httr}.
To bypass Google IP bans, user is strongly encouraged to set the arguments
\code{openvpn} and \code{user_agent} as \code{TRUE}. In this way, every 10-20
sub-requests your IP and your user agent will be automatically and randomly changed.
}
\examples{
\dontrun{
scrap_gscholar()
}
}
\author{
Nicolas CASAJUS, \email{nicolas.casajus@gmail.com}
}
